requirements.txt
# Core
python-dotenv>=1.0
requests>=2.31

# LangChain + GigaChat
langchain>=0.2
langchain-core>=0.2
langchain-gigachat>=0.1.8

# Data / ML
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
joblib>=1.3

# Reporting
openpyxl>=3.1

# (Опционально, но у тебя есть импорты в ноутбуке)
matplotlib>=3.8
seaborn>=0.13

# (Опционально: токенизация для некоторых LLM-ов)
tiktoken>=0.7